Add files via upload
The objective of this project is to conduct an Exploratory Data Analysis (EDA)
on a dataset of Olympic medals. The analysis aims to uncover patterns and insights regarding 
the distribution of medals among different countries, identify trends in medal counts over time, 
and examine the performance of various nations across different Olympic events. 
This EDA will serve as a foundational step for any further analysis or predictive modeling related to 
Olympic performances.